[feat] 페이지네이션과 테이블 선택 로직을 커스텀 훅으로 분리 #211
